--- app/helpers/index.html
@@ -1,50 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Xat</title>
-</head>
-<body>
-    <ul id="messages"></ul>
-    <form id="form" action="">
-      <input id="input" autocomplete="off" /><button>Send</button>
-    </form>
-    <script src="/socket.io/socket.io.js"></script>
-    <script> 
-    var socket = io(); 
-
-    var form = document.getElementById('form');
-    var input = document.getElementById('input');
-    
-    form.addEventListener('submit', function(event) {
-        event.preventDefault(); // The preventDefault() method of the Event interface tells the user agent that if the event does not get explicitly handled, its default action should not be taken as it normally would be.
-        if (input.value) {
-            socket.emit('chat message', input.value);
-            input.value = '';
-        }
-    })
-    
-    socket.on('chat message', function(message) {
-    var item = document.createElement('li'); // document.createElement() method creates the HTML element specified by tagName, or an HTMLUnknownElement if tagName isn't recognized.
-    item.textContent = message;
-    messages.appendChild(item);
-    window.scrollTo(0, document.body.scrollHeight);
-    });
-
-    let count = 0
-    setInterval(() => {
-        socket.volatile.emit('ping', ++count)
-    }, 5000)
-
-    document.addEventListener('keydown', e => {
-        if(e.target.matches('input')) return
-
-        if(e.key === 'c') socket.connect()
-        if (e.key === 'd') socket.disconnect()
-    })
-
-    </script>
-</body>
-</html>
