+++ app/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Xat</title>
+  <script type = "module" src="app.js"></script>
+</head>
+<body>
+  <div id = "message-container"></div>
+  <form id="form">
+    <input id= "message-input" type = "text">
+    <button>Send</button>
+  </form>
+</body>
+</html>
